requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-propagator-gcp==1.11.0
protobuf==6.33.5
alembic==1.14.1
asyncpg==0.30.0
SQLAlchemy==2.0.37
sqlmodel==0.0.22
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PyYAML==6.0.3
requests==2.33.0
requests-oauthlib==2.0.0
rsa==4.9.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1; sys_platform != "win32"
watchfiles==1.1.1
websockets==16.0
